rotation works correctly, identified logic error to fix
fixed rotation issues with the image cycling by creating a global variable numberPlayers to set numPlayers equal to it so the modulo code works outside the generatePlayers function. logic errror is when the most recent bet is set equal to the players most recent bet therefore fulfilling the if statement prematurely, this can be fixed by checking the bigBlind  most recent bet instead i think.